Attempt to stabilize tests
